Move hooks to composables and refactor usage
